servicio2/requirements.txt
Flask
openai==0.28.0
pymongo
PyPDF2
flask-cors
pdfminer.six